Ksvm

+++ vowpalwabbit/kernel_svm.cc
@@ -381,9 +381,9 @@ namespace KSVM
   float kernel_function(const flat_example* fec1, const flat_example* fec2, void* params, size_t kernel_type) {
     switch(kernel_type) {
     case SVM_KER_RBF:
-      return rbf_kernel(fec1, fec2, *((double*)params));
+      return rbf_kernel(fec1, fec2, *((float*)params));
     case SVM_KER_POLY:
-      return poly_kernel(fec1, fec2, *((double*)params));
+      return poly_kernel(fec1, fec2, *((int*)params));
     case SVM_KER_LIN:
       return linear_kernel(fec1, fec2);
     }
@@ -653,11 +653,11 @@ namespace KSVM
       
     if(params.active) {           
       if(params.active_pool_greedy) { 
-	multimap<double, int> scoremap;
+	multimap<double, size_t> scoremap;
 	for(size_t i = 0;i < params.pool_pos; i++)
-	  scoremap.insert(pair<const double, const int>(fabs(scores[i]),i));
+	  scoremap.insert(pair<const double, const size_t>(fabs(scores[i]),i));
 
-	multimap<double, int>::iterator iter = scoremap.begin();
+	multimap<double, size_t>::iterator iter = scoremap.begin();
 	//cerr<<params.pool_size<<" "<<"Scoremap: ";
 	//for(;iter != scoremap.end();iter++)
 	//cerr<<iter->first<<" "<<iter->second<<" "<<((label_data*)params.pool[iter->second]->ld)->label<<"\t";
@@ -855,7 +855,7 @@ namespace KSVM
       ("pool_size", po::value<size_t>(), "size of pools for active learning")
       ("subsample", po::value<size_t>(), "number of items to subsample from the pool")
       ("kernel", po::value<string>(), "type of kernel (rbf or linear (default))")
-      ("bandwidth", po::value<double>(), "bandwidth of rbf kernel")
+      ("bandwidth", po::value<float>(), "bandwidth of rbf kernel")
       ("degree", po::value<int>(), "degree of poly kernel")
       ("lambda", po::value<double>(), "saving regularization for test time");
 
@@ -919,16 +919,16 @@ namespace KSVM
 
     if(kernel_type.compare("rbf") == 0) {
       params->kernel_type = SVM_KER_RBF;
-      double bandwidth = 1.;
+      float bandwidth = 1.;
       if(vm.count("bandwidth")) {
-	std::stringstream ss;
-	bandwidth = vm["bandwidth"].as<double>();	
-	ss<<" --bandwidth "<<bandwidth;
-	all.file_options.append(ss.str());
+		std::stringstream ss;
+		bandwidth = vm["bandwidth"].as<float>();	
+		ss<<" --bandwidth "<<bandwidth;
+		all.file_options.append(ss.str());
       }
       cerr<<"bandwidth = "<<bandwidth<<endl;
       params->kernel_params = calloc(1,sizeof(double*));
-      *((double*)params->kernel_params) = bandwidth;
+      *((float*)params->kernel_params) = bandwidth;
     }
     else if(kernel_type.compare("poly") == 0) {
       params->kernel_type = SVM_KER_POLY;
@@ -940,12 +940,12 @@ namespace KSVM
 	  all.file_options.append(ss.str());
 	}
       cerr<<"degree = "<<degree<<endl;
-      params->kernel_params = calloc(1,sizeof(double*));
-      *((double*)params->kernel_params) = degree;
+      params->kernel_params = calloc(1,sizeof(int*));
+      *((int*)params->kernel_params) = degree;
     }      
     else
       params->kernel_type = SVM_KER_LIN;            
-
+	
     all.reg.weight_mask = (uint32_t)FLT_MAX;
     params->all->reg.weight_mask = (uint32_t)FLT_MAX;
